📝 userInfo -> profile 폴더 이름 변경
