+++ src/app/(main)/profile/edit-account/page.tsx
@@ -1,0 +1,5 @@
+import EditAccount from '@/widgets/profile/edit-account/EditAccount';
+
+export default function EditAccountPage() {
+  return <EditAccount />;
+}
