+++ src/app/(main)/profile/edit-profile/page.tsx
@@ -1,0 +1,5 @@
+import EditProfile from '@/widgets/profile/edit-profile/EditProfile';
+
+export default function EditProfilePage() {
+  return <EditProfile />;
+}
