--- src/app/(main)/user/edit-account/page.tsx
@@ -1,5 +1,0 @@
-import EditAccount from "@/widgets/userInfo/edit-account/EditAccount";
-
-export default function EditAccountPage() {
-  return <EditAccount />
-}
