--- src/app/(main)/user/edit-profile/page.tsx
@@ -1,5 +1,0 @@
-import EditProfile from "@/widgets/userInfo/edit-profile/EditProfile";
-
-export default function EditProfilePage() {
-  return <EditProfile />
-}
