+++ src/widgets/profile/CategoryTabs.tsx
@@ -1,4 +1,4 @@
-import { userInfoStore } from '@/entities/userInfo/UserInfoStore';
+import { userInfoStore } from '@/entities/profile/UserInfoStore';
 import { colorChangeAnimation } from '@/shared/ui/Animation';
 import { useState } from 'react';
