+++ src/widgets/profile/edit-profile/Types.ts
@@ -1,4 +1,4 @@
-import { step2Schema } from '@/features/userInfo/model/Step2';
+import { step2Schema } from '@/features/profile/model/Step2';
 import { step3Schema } from '@/features/auth/signup/model/schema/Step3';
 import { step4Schema } from '@/features/auth/signup/model/schema/Step4';
 import { Dispatch, JSX, SetStateAction } from 'react';
